requirements.txt
fastapi==0.116.1
uvicorn==0.35.0
python-multipart==0.0.20
pydantic==2.11.7
pydantic-settings==2.10.1
python-docx==1.2.0
email-validator==2.2.0
pytest==8.3.4
Pillow==11.1.0
transformers==4.44.2
huggingface_hub==0.24.7
torch==2.4.1
safetensors==0.4.5
SQLAlchemy==2.0.36
psycopg2-binary==2.9.10